Refactor: modify profile component to profile screen
- Profile 파일 이름 및 컴포넌트명 수정
